+++ crates/fuzz/src/fuzz_data.rs
@@ -11,6 +11,7 @@ use std::collections::HashMap;
 use std::error::Error;
 use std::fmt::Display;
 
+use crate::config::Config;
 use crate::error::*;
 use crate::fuzz_client::FuzzClient;
 use crate::fuzz_test_executor::FuzzTestExecutor;
@@ -56,6 +57,7 @@ where
     pub fn run_with_runtime(
         &self,
         client: &mut impl FuzzClient,
+        config: &Config,
     ) -> core::result::Result<(), Box<dyn Error + 'static>> {
         // solana_logger::setup_with_default("off");
         // #[cfg(fuzzing_debug)]
@@ -83,7 +85,7 @@ where
             eprintln!("\x1b[34mCurrently processing\x1b[0m: {}", fuzz_ix);
 
             if fuzz_ix
-                .run_fuzzer(&self.accounts, client, &mut sent_txs)
+                .run_fuzzer(&self.accounts, client, &mut sent_txs, config)
                 .is_err()
             {
                 // for now skip following instructions in case of error and move to the next fuzz iteration